refactor: keep same structure with huggingface except the `mindone` prefix
