+++ mindone/diffusers/configuration_utils.py
