+++ mindone/diffusers/image_processor.py
