+++ mindone/diffusers/models/modeling_outputs.py
