+++ mindone/diffusers/models/resnet.py
