+++ mindone/diffusers/pipelines/stable_diffusion_xl/__init__.py
