+++ mindone/diffusers/pipelines/stable_diffusion_xl/pipeline_output.py
