+++ mindone/diffusers/utils/hub_utils.py
