+++ mindone/diffusers/utils/outputs.py
