+++ mindone/transformers/models/clip/__init__.py
