+++ mindone/transformers/models/clip/modeling_ms_clip.py
